Deb: Trash Management Initial Work -->
At the start of the game, 3 factories are automatically created. The mutant
shark is a organic factory, orca is the plastic factory and the huge
terminator is the nuclear factory.
SUSHIL -> remove this when factory creation menu is in place. Search your
name is the code.
Also, replace the factory models when the real models are ready.

Right click of factory to open the configuation screen. Nothing is there at
present.

Collect different trash, and double click on factory to dump them.

Power packs are not generated yet.